moved components from home to app to apply header and footer to all pages, isolated search into its own component (does not affect layout)
