+++ src/App.js
@@ -4,26 +4,40 @@ import { BrowserRouter as Router, Switch, Route, useParams} from "react-router-d
 import Home from './components/pages/Home';
 import Shop from './components/pages/Shop';
 import About from './components/pages/About';
-import SearchIcon from "@material-ui/icons/Search";
-import Search from './components/pages/Search/Search';
+import Footer from "./components/Footers/Footer"
+import BottomFooter from "./components/Footers/BottomFooter"
+import NavBar from "./components/NavBar"
+import SocialBar from "./components/socialsbar/socialsbar"
+import Result from './components/pages/Result/Result'
 
 
 const App = () => {
   // const [dropdown, setDropdown] = useState(false)
+  const [search, setSearch] = useState("");
+  const [searchResults, setSearchResults] = useState([]);
+
+  const searchHandler = (val) => {
+    console.log(val);
+    setSearch(val.target.value);
+  } 
  
   
     return (
       <div className="App">
         <Router>
+        <NavBar />
+            <SocialBar />
           <div className="App-header">
             <Switch>
               <Route path='/' exact component={Home}  />
               <Route path='/Shop' exact component={Shop}  />
               <Route path='/About' exact component={About}  />
-              {/* <Route path='/search/' exact component={Search} /> */}
+              <Route path='/search/' exact component={Result}/>
             </Switch>
           </div>
         </Router>
+        <Footer />
+            <BottomFooter />
       </div>
     );
 };
