+++ src/components/pages/Home.js
@@ -6,16 +6,10 @@ import unionjacklargecushion from "../images/unionjacklargecushion.jpg"
 import glasses from "../images/glasses.jpg"
 import flipper from "../images/flipper.png"
 import Featuredproducts from "../products/featuredproducts"
-import NavBar from "../NavBar" 
-import Footer from "../Footers/Footer"
-import BottomFooter from "../Footers/BottomFooter"
-import SocialBar from "../socialsbar/socialsbar";
 
 function Home() {
     return (
         <div>
-            <NavBar />
-            <SocialBar />
             <div className="homecontainer">
                 <MainCaro/>
             <div className="newarrivals">
@@ -26,8 +20,7 @@ function Home() {
                     <Featuredproducts featuredimage={flipper}/>
                  </div>
             </div>
-            <Footer />
-            <BottomFooter />
+
         </div>  
         </div>
