+++ src/components/pages/Result/Result.js
@@ -1,0 +1,13 @@
+import react from 'react';
+
+const Result = () => {
+
+    return (
+        <div>
+            
+        </div>
+        
+    )
+}
+
+export default Result;
