+++ src/components/pages/Search/Search.js
@@ -1,15 +1,34 @@
-import react from 'react';
-import {useParams, useLocation} from 'react-router-dom';
+import react, {useState} from 'react';
+import SearchIcon from "@material-ui/icons/Search";
 
 
 
 const Search = () => {
     
+    const [search, setSearch] = useState("");
+    const [searchResults, setSearchResults] = useState([]);
+
+    const searchHandler = (val) => {
+        setSearch(val.target.value);
+      } 
+    
+    const dataSearch = async () => {
+    
+    const res = await fetch(`http://localhost:5000/search/${search}`);
+    
+    const data = await res.json();
+    setSearchResults(data);
+    console.log(data);
+    }
+
     return (
-        <div>
-            <p>Search:</p>
+        <div className="left_header">
+        <SearchIcon className="search" onClick={dataSearch}/>
+        {/* <Link to="/result"><SearchIcon className="search"/></Link> */}
+                <input placeholder="Search products" className="search"
+                type="text" onChange={(e)=>searchHandler(e)}/>
         </div>
-        
+        //onChange={(e)=>searchHandler(e)}
     )
 }
